destructure RecordDetails.js & UserRecordItem.js with component CopyLinkToClipboard.js
