+++ frontend/src/components/records/UserRecordItem.js
@@ -1,10 +1,9 @@
-import React, {useState} from 'react'
+import React from 'react'
 import styled from 'styled-components/macro';
-import {FaLink} from "react-icons/fa";
 import UserTagList from "../tags/UserTagList";
+import CopyLinkToClipboard from "../CopyLinkToClipboard";
 
 export default function UserRecordItem({record, actions = [], className}) {
-    const [copySuccess, setCopySuccess] = useState('');
 
     return (
         <SingleRecordStyled className={className}>
@@ -13,34 +12,10 @@ export default function UserRecordItem({record, actions = [], className}) {
             <UserTagList tags={record.tagList}/>
             <ActionSection>
                 <div>{actions}</div>
-                <CopyLinkStyled>
-                    <CopyLinkIconStyled onClick={copyToClipBoard}/>
-                    {copySuccess}
-                </CopyLinkStyled>
+                <CopyLinkToClipboard recordLink={record.recordLink} />
             </ActionSection>
         </SingleRecordStyled>
-
     )
-
-    function copyToClipBoard(e) {
-        const textArea = document.createElement("textarea")
-        textArea.value = record.recordLink
-        document.body.appendChild(textArea);
-        document.execCommand('copy');
-        e.target.focus();
-        textArea.select();
-        setCopySuccess('Copied!');
-
-        try {
-            const successful = document.execCommand('copy');
-            const msg = successful ? 'successful!' : 'unsuccessful!';
-            setCopySuccess(msg);}
-        catch (err){
-            setCopySuccess('Opps, unable to copy');}
-
-        document.queryCommandSupported('copy')
-        document.body.removeChild(textArea);
-    }
 
 }
 
@@ -58,23 +33,8 @@ justify-items: center;
 align-items: start;
 `
 
-const CopyLinkStyled = styled.div` 
-display: grid;
-grid-template-rows: min-content min-content;
-justify-items: center;
-align-items: center;
-text-align: center;
-color: var(--grey-50);
-`
-
 const TitelStyled = styled.h2`
 margin: 0;
 color: var(--grey-main);
 text-align: center;
-`
-
-const CopyLinkIconStyled = styled(FaLink)`
-color: var(--secondary1);
-height: 30px;
-width: 30px;
 `
